Refactor CasesCohortButton and SelectCohortsModal to use getCaseIdsFromFilter utility function
* Removed redundant inline getCaseIdsFromFilter logic from both components.
* Centralized case ID retrieval logic in the utils file for better maintainability and reusability.
* Updated imports in both components to utilize the new utility function, enhancing code clarity.
