+++ src/features/cases/CasesCohortButton/utils.ts
@@ -10,3 +10,21 @@ export const getStringOperands = (
 
   return operands;
 };
+
+export const getCaseIdsFromFilter = (
+  filter: any,
+): ReadonlyArray<string> | null => {
+  // Check if filter only contains cases.case_id
+  const root = filter?.root || {};
+  const rootKeys = Object.keys(root);
+  const operands = root['cases.case_id']?.operands;
+  const stringOperands = getStringOperands(operands);
+  if (
+    rootKeys.length === 1 &&
+    rootKeys[0] === 'cases.case_id' &&
+    stringOperands
+  ) {
+    return stringOperands;
+  }
+  return null;
+};
